Feat: "추가 route설정"

+++ src/components/DocumentPage.js
@@ -1,7 +1,8 @@
 import { request } from "../utils/api.js";
+import { push } from "../utils/router.js";
 import DocumentList from "./DocumentList.js";
 
-export default function DocumentPage({ $target }) {
+export default function DocumentPage({ $target, onClickTitle }) {
   const $page = document.createElement("div");
 
   const documentList = new DocumentList({
@@ -15,12 +16,27 @@ export default function DocumentPage({ $target }) {
           parent: "null",
         }),
       });
+      push(`/documents/${createPost.id}`);
     },
+
     onClickRemove: async (id) => {
-      const removePost = await request(`/documents/${id}`, {
-        method: "DELETE",
-      });
+      if (id == window.location.pathname.split("/")[2]) {
+        if (confirm("현재 페이지를 삭제하겠습니까?")) {
+          await request(`/documents/${id}`, {
+            method: "DELETE",
+          });
+          push("/");
+        }
+      } else {
+        if (confirm("해당 페이지를 삭제하겠습니까?")) {
+          await request(`/documents/${id}`, {
+            method: "DELETE",
+          });
+          this.fetchDocument();
+        }
+      }
     },
+
     onClickAdd: async (id) => {
       const addPost = await request(`/documents/`, {
         method: "POST",
@@ -29,10 +45,11 @@ export default function DocumentPage({ $target }) {
           parent: id,
         }),
       });
+      push(`/documents/${addPost.id}`);
     },
   });
 
-  this.setState = async () => {
+  this.fetchDocument = async () => {
     const posts = await request("/documents");
     documentList.setState(posts);
     this.render();
@@ -41,4 +58,15 @@ export default function DocumentPage({ $target }) {
   this.render = () => {
     $target.appendChild($page);
   };
+
+  $page.addEventListener("click", (e) => {
+    const { target } = e;
+    const element = target.closest("li");
+    if (element) {
+      const { id } = element.dataset;
+      if (target.className === "list-title") {
+        onClickTitle(id);
+      }
+    }
+  });
 }
